requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=2.0.0

# Data Processing & File Handling
pandas>=1.3.0
fpdf>=1.7.2
requests>=2.26.0

# AI and Vector Storage
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-readers-web>=0.1.0
openai>=1.0.0

# Async and Scheduling
apscheduler>=3.9.1

# Environment and Configuration
python-dotenv>=0.19.0
